Couple of updates after a long hiatus
I picked this script back up with the intention of repurposing it to
spit out the geojson for all the images. It still does the image
processing (albiet at a smaller size), but now we don't worry about
writing the geojson to couchdb.

I also cleaned up some sloppy bits of code.